The proposed change fixes TypeError: newToast.scope.init is not a function Error
The type error is caused after the first toast is created upon the second and subsequent toasts being created. It occurs because you are not guaranteed that the directive link function has not been fully executed creating the scope.init function. 

By using the $timeout,you are ensuring that the call to use scope.init will occur after the previously queued digest requests have been completed. 

This was verified in angular 1.5.3. and also is not necessary related to the template caching as we repro this using the builtin tpls file.